requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pinecone-client>=3.0.0
redis>=5.0.0
supabase>=2.0.0
sentence-transformers>=2.2.0
whisper>=1.0.0
elevenlabs>=0.2.0